+++ langchain/utils.py
@@ -1,6 +1,6 @@
 """Generic utility functions."""
 import os
-from typing import Any, Dict, List, Optional, Tuple
+from typing import Any, Dict, Optional
 
 
 def get_from_dict_or_env(
@@ -19,19 +19,3 @@ def get_from_dict_or_env(
             f" `{env_key}` which contains it, or pass"
             f"  `{key}` as a named parameter."
         )
-
-
-def register(key: str, _registry: Dict[str, Tuple[Any, List[str]]]) -> Any:
-    """Add a class/function to a registry with required keyword arguments.
-
-    ``_registry`` is a dictionary mapping from a key to a tuple of the class/function
-    and a list of required keyword arguments.
-    """
-
-    def _register_cls(cls: Any, required_kwargs: List) -> Any:
-        if key in _registry:
-            raise ValueError(f"{_registry[key][0]} already registered as {key}")
-        _registry[key] = (cls, required_kwargs)
-        return cls
-
-    return _register_cls
